// packages/playground/src/endpoints/mock-server/data/api/types.ts
export enum TaskTypes {
    BUSINESS_TYPE_SELECTION = 'BUSINESS_TYPE_SELECTION',
    INTRODUCTION = 'INTRODUCTION',
    TASKS_OVERVIEW = 'TASKS_OVERVIEW',
    REVIEW = 'REVIEW',
    INDIVIDUAL = 'INDIVIDUAL',
    COMPANY = 'COMPANY',
    COMPANY_SEARCH = 'COMPANY_SEARCH',
    PAYOUT = 'PAYOUT',
    PAYIN = 'PAYIN',
    SERVICE_AGREEMENT = 'SERVICE_AGREEMENT',
    PCI_DSS = 'PCI_DSS',
    DECISION_MAKER = 'DECISION_MAKER',
    DECISION_MAKER_OVERVIEW = 'DECISION_MAKER_OVERVIEW',
    TRUST = 'TRUST',
    TRUST_MEMBER_OVERVIEW = 'TRUST_MEMBER',
    TRUST_MEMBER_INDIVIDUAL = 'TRUST_MEMBER_INDIVIDUAL',
    TRUST_MEMBER_COMPANY = 'TRUST_MEMBER_COMPANY',
    TRUST_MEMBER_ROLE_AND_TYPE = 'TRUST_MEMBER_ROLE_AND_TYPE',
    TRUST_MEMBER_COMPANY_OWNER = 'TRUST_MEMBER_COMPANY_OWNER',
    SOLE_PROPRIETOR_COMPANY = 'SOLE_PROPRIETOR_COMPANY',
    LEGAL_ENTITY_TYPE_SWITCHER = 'LEGAL_ENTITY_TYPE_SWITCHER',
    CAPABILITY_REJECTED = 'CAPABILITY_REJECTED',
}